tmt: Try to describe how to reproduce tmt+packages locally
The whole chain of packit -> tmt+testing-farm is very complicated
and I struggle to wrap my head around it even *without* involving
bootc.

With a bit of AI assistance we came up with this, which
I didn't try yet truly running and clearly if we were to do
so we'd need to automate it.

And doing *that* really is probably going to pull us to
split out "build an rpm" from "build a container" because
the former is what the packit flow wants today.

xref https://github.com/teemtee/tmt/issues/1018

Signed-off-by: Colin Walters <walters@verbum.org>