tmt: Remove one level of test enumeration duplication
Previously we duplicated all test names *three* times:

- In the plan fmf
- In the test fmf
- In the test binary itself

Now while we could be using plans better, because we're provisioning
with bcvk now we have to basically bypass the plan stuff and
directly run a test.

So use the plans for logical grouping, but change our runner
code to just directly run the tests.

Signed-off-by: Colin Walters <walters@verbum.org>